Base timezones in UTC
After installing this gem on my site, I noticed that the timing of
orders was off on mailchimp's UI.  I discovered that even if you
annotate your timezone, Mailchimp assumes the times you send in iso8601
format are in UTC.  Casting all times to UTC (since `completed_at` etc.
use locally time-zoned times) and then using the `iso8601` helper
simplified this code and results in the correct time appearing in the
mailchimp console.