Fix nil error in order.mailchimp_cart
In my historical dataset of spree orders, there are some promotion
actions that don't actually have a source.  The first `map` in this line
can return `[nil]` if the only adjustment/promotion (for whatever
reason) doesn't have a source.  In these cases, the `&:promotion` in the
second map will fail.  To fix, just `compact` out the nil sources.