fix codeless promos
In spree, it's valid to have a promotion without a code (e.g. automatic
free shipping promos often have this).  In mailchimp, a promo cannot
have an empty string as a code which is what was defaulted to in the
promotion presenters.  Instead, default to a placeholder string:
`"promotion:#{promotion.id}"`.